refactored func SetContainerTemplateRelation from lxd package to cmd package corrected typo: "relations"
